front parece todo ok con la nueva gestion de session

+++ frontend/restrict/src/modules/login/hooks/custom_login.js
@@ -1,7 +1,7 @@
 import {useEffect} from "react"
 import {useHistory} from "react-router-dom"
 import db from "helpers/localdb"
-import {async_get_user_by_tpvcode, async_insert_rnd} from "modules/login/async/async_repository"
+import {async_get_session_id, async_get_user_by_tpvcode, async_insert_rnd} from "modules/login/async/async_repository"
 
 function CustomLogin() {
   const history = useHistory()
@@ -9,14 +9,18 @@ function CustomLogin() {
   const on_submit = async tpvcode => {
     const pincode = tpvcode.toString()
     let r = await async_get_user_by_tpvcode(pincode)
-    console.log("result r tpv", r)
-    db.delete("user_session")
+
+    db.delete("session_user")
+    db.delete("session_id")
     if(parseInt(r.foundrows)===1) {
-      console.log("async_get_user_by_tpvcode",r)
+
       const token = db.select("token_apify")
       const usermini = r.result[0]
       await async_insert_rnd(token, usermini)
-      db.save("user_session", usermini.code_cache)
+      r = await async_get_session_id(token, usermini?.code_cache)
+      if(!r) return
+      db.save("session_user", usermini)
+      db.save("session_id", r)
 
       let url = db.select("last_location")
       db.delete("last_location")
@@ -28,7 +32,8 @@ function CustomLogin() {
   }
 
   useEffect(() => {
-    db.delete("user_session")
+    db.delete("session_user")
+    db.delete("session_id")
     return () => console.log("login.insert.unmounting")
   },[])
